Fix bug related to syncing incomplete external data to ftp
Getting all files from the database does not work because files are
being added to the db while they’ve been synced to the staging area. It
is better to rely with files on disk instead after all read 1 fastq
files are in db.
Added a new test folder for testing this change.